+++ examples/postgresflex/postgresflex.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/stackitcloud/stackit-sdk-go/core/config"
 	"github.com/stackitcloud/stackit-sdk-go/services/postgresflex"
+	"github.com/stackitcloud/stackit-sdk-go/services/postgresflex/wait"
 )
 
 func main() {
@@ -50,4 +51,33 @@ func main() {
 	}
 
 	fmt.Printf("Created user \"%s\" associated to instance \"%s\".\n", username, *items[0].Name)
+
+	// Delete an instance
+	err = postgresflexClient.DeleteInstance(context.Background(), projectId, instanceId).Execute()
+
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error when delete PostgreSQL Flex instance: %v", err)
+	}
+
+	_, err = wait.DeleteInstanceWaitHandler(context.Background(), postgresflexClient, projectId, instanceId).WaitWithContext(context.Background())
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error when waiting for PostgreSQL Flex instance deletion: %v", err)
+	}
+
+	fmt.Printf("Deleted PostgreSQL Flex instance \"%s\".\n", instanceId)
+
+	// Force delete an instance
+	err = postgresflexClient.ForceDeleteInstance(context.Background(), projectId, instanceId).Execute()
+
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error when force delete PostgreSQL Flex instance: %v", err)
+	}
+
+	_, err = wait.ForceDeleteInstanceWaitHandler(context.Background(), postgresflexClient, projectId, instanceId).WaitWithContext(context.Background())
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error when waiting for PostgreSQL Flex instance force deletion: %v", err)
+	}
+
+	fmt.Printf("Force deleted PostgreSQL Flex instance \"%s\".\n", instanceId)
+
 }